Added support for eject button. Cleaned up a few files.
