+++ src/control/display_state.cpp
@@ -23,166 +23,6 @@
 
 using namespace zuluide::control;
 
-StatusState::StatusState (int imgNameOffset) : imageNameOffset(imgNameOffset) { }
-
-StatusState::StatusState (const StatusState& src) : imageNameOffset(src.imageNameOffset) { }
-
-int StatusState::GetImageNameOffset () const {
-  return imageNameOffset;
-}
-
-void StatusState::IncrementImageNameOffset() {
-  imageNameOffset++;
-}
-
-void StatusState::DecrementImageNameOffset() {
-  imageNameOffset--;
-  if (imageNameOffset < 0) {
-    imageNameOffset = 0;
-  }
-}
-
-void StatusState::ResetImageNameOffset() {
-  imageNameOffset = 0;
-}
-
-StatusState& StatusState::operator=(const StatusState& src) {
-  imageNameOffset = src.imageNameOffset;
-  return *this;
-}
-
-MenuState::MenuState (MenuState::Entry value) : currentEntry(value) { }
-
-MenuState::MenuState (const MenuState &src) : currentEntry(src.currentEntry) { }
-
-MenuState::Entry MenuState::GetCurrentEntry () const {
-  return currentEntry;
-}
-
-MenuState& MenuState::operator=(const MenuState& src) {
-  currentEntry = src.currentEntry;
-  return *this;
-}
-
-void MenuState::MoveToNextEntry() {
-  switch (currentEntry) {
-  case Entry::Eject: {
-    currentEntry = Entry::New;
-    break;
-  }
-    
-  case Entry::Select: {
-    currentEntry = Entry::Eject;
-    break;
-  }
-    
-  case Entry::New: {
-    currentEntry = Entry::Back;
-    break;
-  }
-    
-  case Entry::Back: {
-    currentEntry = Entry::Select;
-    break;
-  }
-  }
-}
-
-void MenuState::MoveToPreviousEntry() {
-  switch (currentEntry) {
-  case Entry::Eject: {
-    currentEntry = Entry::Back;
-    break;
-  }
-    
-  case Entry::Select: {
-    currentEntry = Entry::Eject;
-    break;
-  }
-    
-  case Entry::New: {
-    currentEntry = Entry::Select;
-    break;
-  }
-    
-  case Entry::Back: {
-    currentEntry = Entry::New;
-    break;
-  }
-  }
-}
-
-SelectState::SelectState (int imgNameOffset) : imageNameOffset(imgNameOffset), isShowingBack(false) {}
-
-SelectState::SelectState (const SelectState& src) : imageNameOffset(src.imageNameOffset), isShowingBack(src.isShowingBack) {
-  if (src.currentImage) {
-    currentImage = std::make_unique<Image>(*src.currentImage);
-  }
-}
-
-int SelectState::GetImageNameOffset () const {
-  return imageNameOffset;
-}
-
-SelectState& SelectState::operator=(const SelectState& src) {
-  imageNameOffset = src.imageNameOffset;
-  isShowingBack = src.isShowingBack;
-  if (src.currentImage) {
-    currentImage = std::make_unique<Image>(*src.currentImage);
-  } else {
-    currentImage = nullptr;
-  }
-  
-  return *this;
-}
-
-void SelectState::SetImageNameOffset(int value) {
-  imageNameOffset = value;
-}
-
-void SelectState::SetCurrentImage(std::unique_ptr<zuluide::images::Image> image) {
-  currentImage = std::move(image);
-}
-
-const zuluide::images::Image& SelectState::GetCurrentImage() const {
-  return *currentImage;
-}
-
-bool SelectState::HasCurrentImage() const {
-  return currentImage ? true : false;
-}
-
-bool SelectState::IsShowingBack() const {
-  return isShowingBack;
-}
-
-void SelectState::SetIsShowingBack(bool value) {
-  isShowingBack = value;
-}
-
-NewImageState::NewImageState (int imgIndex) : imageIndex (imgIndex) {}
-
-NewImageState::NewImageState (const NewImageState& src) : imageIndex(src.imageIndex) {}
-
-int NewImageState::GetImageIndex () const {
-  return imageIndex;
-}
-
-NewImageState& NewImageState::operator=(const NewImageState& src) {
-  imageIndex = src.imageIndex;
-  return *this;
-}
-
-NewImageState& NewImageState::operator++(int) {
-  imageIndex++;
-  return *this;
-}
-
-NewImageState& NewImageState::operator--(int) {
-  imageIndex--;
-  return *this;
-}
-
 DisplayState::DisplayState(const StatusState &state) :
   currentMode(Mode::Status),
   statusState(state)
